Show feed descriptions in a resizable popup, not in teeny tiny box.
